Update build-multiarch.sh
Changed conditional statements, should now build for non-arm as well.

+++ build-multiarch.sh
@@ -17,32 +17,36 @@ if [[ ! -z "$DOCKERFILE_PATH" && ! -z "$TAG" ]] ; then
 	BUILD_ARCH=$(echo "${DOCKERFILE_PATH}" | cut -d '.' -f 2)
 
 	# For amd64 filename is Dockerfile
-	[[ "${BUILD_ARCH}" == "Dockerfile" || "${BUILD_ARCH}" == "i386" ]] && \
-		{ echo 'qemu-user-static: Download not required for current arch'; exit 0; }
+	if [[ "${BUILD_ARCH}" != "Dockerfile" || "${BUILD_ARCH}" != "i386" ]] ; then
 
-	case ${BUILD_ARCH} in
-		amd64 ) QEMU_ARCH="x86_64" ;;
-		i386 ) QEMU_ARCH="i386" ;;
-		arm32v6 ) QEMU_ARCH="arm" ;;
-		arm32v7 ) QEMU_ARCH="arm" ;;
-		arm64v8 ) QEMU_ARCH="aarch64" ;
-	esac
+		case ${BUILD_ARCH} in
+			amd64 ) QEMU_ARCH="x86_64" ;;
+			i386 ) QEMU_ARCH="i386" ;;
+			arm32v6 ) QEMU_ARCH="arm" ;;
+			arm32v7 ) QEMU_ARCH="arm" ;;
+			arm64v8 ) QEMU_ARCH="aarch64" ;
+		esac
 
-	QEMU_USER_STATIC_DOWNLOAD_URL="https://github.com/multiarch/qemu-user-static/releases/download"
-	QEMU_USER_STATIC_LATEST_TAG=$(curl -s https://api.github.com/repos/multiarch/qemu-user-static/tags \
-		| grep 'name.*v[0-9]' \
-		| head -n 1 \
-		| cut -d '"' -f 4)
+		if [[ -z QEMU_ARCH ]] ; then
+			echo "Unsupported architecture" && exit(0);
+		fi
 
-	curl -SL "${QEMU_USER_STATIC_DOWNLOAD_URL}/${QEMU_USER_STATIC_LATEST_TAG}/x86_64_qemu-${QEMU_ARCH}-static.tar.gz" \
-		| tar xzv
+		QEMU_USER_STATIC_DOWNLOAD_URL="https://github.com/multiarch/qemu-user-static/releases/download"
+		QEMU_USER_STATIC_LATEST_TAG=$(curl -s https://api.github.com/repos/multiarch/qemu-user-static/tags \
+			| grep 'name.*v[0-9]' \
+			| head -n 1 \
+			| cut -d '"' -f 4)
+
+		curl -SL "${QEMU_USER_STATIC_DOWNLOAD_URL}/${QEMU_USER_STATIC_LATEST_TAG}/x86_64_qemu-${QEMU_ARCH}-static.tar.gz" \
+			| tar xzv
+			
+		BUILD_ARCH=$(echo "${DOCKERFILE_PATH}" | cut -d '.' -f 2)
+
+		docker run --rm --privileged multiarch/qemu-user-static:register --reset
 		
-	BUILD_ARCH=$(echo "${DOCKERFILE_PATH}" | cut -d '.' -f 2)
-
-	[ "${BUILD_ARCH}" == "Dockerfile" ] && \
-		{ echo 'qemu-user-static: Registration not required for current arch'; exit 0; }
-
-	docker run --rm --privileged multiarch/qemu-user-static:register --reset
+	else
+		echo 'qemu-user-static: Download not required for current arch'
+	fi
 	
 	docker build -t ${TAG} -f $DOCKERFILE_PATH .
 else
